Almost done the save of the range in db. TODO: set up correctly the update of the new range in range checker. Build a conversion from int to RangeType
